[FEATURE] #15 - 패키지 수정
