+++ app/src/main/java/co/kr/woowahan_accountbook/di/SettingModule.kt
@@ -2,9 +2,8 @@ package co.kr.woowahan_accountbook.di
 
 import co.kr.woowahan_accountbook.data.datasource.local.classification.ClassificationDataSource
 import co.kr.woowahan_accountbook.data.datasource.local.payment.PaymentDataSource
-import co.kr.woowahan_accountbook.data.repository.SettingRepositoryImpl
+import co.kr.woowahan_accountbook.data.repository.setting.SettingRepositoryImpl
 import co.kr.woowahan_accountbook.domain.repository.setting.SettingRepository
-import co.kr.woowahan_accountbook.domain.usecase.SettingPaymentsUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
