+++ app/src/main/java/co/kr/woowahan_accountbook/domain/usecase/setting/SettingPaymentUpdateUseCase.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_accountbook.domain.usecase
+package co.kr.woowahan_accountbook.domain.usecase.setting
 
 import co.kr.woowahan_accountbook.domain.repository.setting.SettingRepository
 import kotlinx.coroutines.CoroutineDispatcher
