+++ app/src/main/java/co/kr/woowahan_accountbook/presentation/adapter/setting/SettingClassificationAdapter.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_accountbook.presentation.adapter
+package co.kr.woowahan_accountbook.presentation.adapter.setting
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,31 +6,31 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import co.kr.woowahan_accountbook.R
 import co.kr.woowahan_accountbook.databinding.ItemSettingBodyAddBinding
-import co.kr.woowahan_accountbook.databinding.ItemSettingBodyPaymentBinding
+import co.kr.woowahan_accountbook.databinding.ItemSettingBodyClassificationBinding
 import co.kr.woowahan_accountbook.databinding.ItemSettingHeaderBinding
-import co.kr.woowahan_accountbook.domain.entity.setting.SettingPayment
+import co.kr.woowahan_accountbook.domain.entity.setting.SettingClassification
 
-class SettingPaymentAdapter(
-    private val updatePayment: (SettingPayment) -> Unit,
-    private val addPayment: () -> Unit
+class SettingClassificationAdapter(
+    private val onClassificationClick: (SettingClassification) -> Unit,
+    private val onAddButtonClick: () -> Unit
 ) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-    private val items = mutableListOf<SettingPayment>()
+    private val items = mutableListOf<SettingClassification>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return when (viewType) {
-            SettingPayment.Type.HEADER -> HeaderViewHolder.create(parent)
-            SettingPayment.Type.BODY -> BodyViewHolder.create(parent, updatePayment)
-            SettingPayment.Type.FOOTER -> FooterViewHolder.create(parent, addPayment)
+            SettingClassification.Type.HEADER -> HeaderViewHolder.create(parent)
+            SettingClassification.Type.BODY -> BodyViewHolder.create(parent, onClassificationClick)
+            SettingClassification.Type.FOOTER -> FooterViewHolder.create(parent, onAddButtonClick)
             else -> throw IllegalArgumentException()
         }
     }
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         when (holder) {
-            is HeaderViewHolder -> holder.onBind()
+            is HeaderViewHolder -> holder.onBind(items[position].isIncome)
             is BodyViewHolder -> holder.onBind(items[position])
-            is FooterViewHolder -> holder.onBind()
+            is FooterViewHolder -> holder.onBind(items[position].isIncome)
         }
     }
 
@@ -40,8 +40,8 @@ class SettingPaymentAdapter(
 
     class HeaderViewHolder(private val binding: ItemSettingHeaderBinding) :
         RecyclerView.ViewHolder(binding.root) {
-        fun onBind() {
-            binding.tvHeaderTitle.text = "결제수단"
+        fun onBind(type: Boolean) {
+            binding.tvHeaderTitle.text = if (type) "수입 카테고리" else "지출 카테고리"
         }
 
         companion object Factory {
@@ -57,55 +57,58 @@ class SettingPaymentAdapter(
     }
 
     class BodyViewHolder(
-        private val binding: ItemSettingBodyPaymentBinding,
-        private val updatePayment: (SettingPayment) -> Unit
+        private val binding: ItemSettingBodyClassificationBinding,
+        private val onClassificationClick: (SettingClassification) -> Unit
     ) :
         RecyclerView.ViewHolder(binding.root) {
-        fun onBind(payment: SettingPayment) {
-            binding.payment = payment
+        fun onBind(classification: SettingClassification) {
+            binding.classification = classification
             binding.root.setOnClickListener {
-                updatePayment(payment)
+                onClassificationClick(classification)
             }
         }
 
         companion object Factory {
-            fun create(parent: ViewGroup, updatePayment: (SettingPayment) -> Unit): BodyViewHolder {
-                val binding = DataBindingUtil.inflate<ItemSettingBodyPaymentBinding>(
+            fun create(
+                parent: ViewGroup,
+                onClassificationClick: (SettingClassification) -> Unit
+            ): BodyViewHolder {
+                val binding = DataBindingUtil.inflate<ItemSettingBodyClassificationBinding>(
                     LayoutInflater.from(parent.context),
-                    R.layout.item_setting_body_payment,
+                    R.layout.item_setting_body_classification,
                     parent,
                     false
                 )
-                return BodyViewHolder(binding, updatePayment)
+                return BodyViewHolder(binding, onClassificationClick)
             }
         }
     }
 
     class FooterViewHolder(
         private val binding: ItemSettingBodyAddBinding,
-        private val addPayment: () -> Unit
+        private val onAddButtonClick: () -> Unit
     ) :
         RecyclerView.ViewHolder(binding.root) {
-        fun onBind() {
-            binding.tvBodyAdd.text = "결제수단 추가하기"
+        fun onBind(type: Boolean) {
+            binding.tvBodyAdd.text = if (type) "수입 카테고리 추가하기" else "지출 카테고리 추가하기"
             binding.root.setOnClickListener {
-                addPayment()
+                onAddButtonClick()
             }
         }
 
         companion object Factory {
-            fun create(parent: ViewGroup, addPayment: () -> Unit): FooterViewHolder {
+            fun create(parent: ViewGroup, onAddButtonClick: () -> Unit): FooterViewHolder {
                 val binding = ItemSettingBodyAddBinding.inflate(
                     LayoutInflater.from(parent.context),
                     parent,
                     false
                 )
-                return FooterViewHolder(binding, addPayment)
+                return FooterViewHolder(binding, onAddButtonClick)
             }
         }
     }
 
-    fun updateItems(newItems: List<SettingPayment>) {
+    fun updateItems(newItems: List<SettingClassification>) {
         items.clear()
         items.addAll(newItems)
         notifyDataSetChanged()
