+++ app/src/main/java/co/kr/woowahan_accountbook/presentation/viewmodel/main/setting/PaymentAddViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import co.kr.woowahan_accountbook.domain.usecase.SettingPaymentAddUseCase
-import co.kr.woowahan_accountbook.domain.usecase.SettingPaymentUpdateUseCase
+import co.kr.woowahan_accountbook.domain.usecase.setting.SettingPaymentAddUseCase
+import co.kr.woowahan_accountbook.domain.usecase.setting.SettingPaymentUpdateUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import timber.log.Timber
